Added access and refresh token logic using flask-jwt-extended

+++ resources/item.py
@@ -1,13 +1,26 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required,
+    get_jwt_claims,
+    jwt_optional,
+    get_jwt_identity
+)
 
 from models.item import ItemModel
 
 
 class Items(Resource):
 
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'Login to get more information'
+        }
 
 
 class Item(Resource):
@@ -25,7 +38,7 @@ class Item(Resource):
         help="Every item needs a store id"
     )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -44,7 +57,12 @@ class Item(Resource):
             return {'message': 'Failed to add item'}, 500
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Unauthorized access'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
